remove previous generator
